Update to version 3.22

+++ usr.bin/file/is_tar.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: is_tar.c,v 1.6 1997/01/09 20:18:55 tls Exp $	*/
+/*	$NetBSD: is_tar.c,v 1.7 1997/01/28 00:49:42 christos Exp $	*/
 
 /*
  * is_tar() -- figure out whether file is a tar archive.
@@ -7,7 +7,7 @@
  * Pubic Domain version written 26 Aug 1985 John Gilmore (ihnp4!hoptoad!gnu).
  *
  * @(#)list.c 1.18 9/23/86 Public Domain - gnu
- * $NetBSD: is_tar.c,v 1.6 1997/01/09 20:18:55 tls Exp $
+ * $NetBSD: is_tar.c,v 1.7 1997/01/28 00:49:42 christos Exp $
  *
  * Comments changed and some code/comments reformatted
  * for file command by Ian Darwin.
@@ -21,9 +21,9 @@
 #define	isodigit(c)	( ((c) >= '0') && ((c) <= '7') )
 
 #if	defined(__STDC__) || defined(__cplusplus)
-static long from_oct(int, char*);	/* Decode octal number */
+static int from_oct(int, char*);	/* Decode octal number */
 #else
-static long from_oct();
+static int from_oct();
 #endif
 
 /*
@@ -39,7 +39,7 @@ int nbytes;
 {
 	register union record *header = (union record *)buf;
 	register int	i;
-	register long	sum, recsum;
+	register int	sum, recsum;
 	register char	*p;
 
 	if (nbytes < sizeof(union record))
@@ -77,12 +77,12 @@ int nbytes;
  *
  * Result is -1 if the field is invalid (all blank, or nonoctal).
  */
-static long
+static int
 from_oct(digs, where)
 	register int	digs;
 	register char	*where;
 {
-	register long	value;
+	register int	value;
 
 	while (isspace(*where)) {		/* Skip spaces */
 		where++;
